Backend/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.10.5
azure-cosmos==4.3.0
